+++ src/libclient/net/message.h
@@ -91,7 +91,7 @@ Message makeLayerAclMessage(
 	const QVector<uint8_t> &exclusive);
 
 Message makeLayerTreeCreateMessage(
-	uint8_t contextId, uint32_t id, uint16_t source, uint16_t target,
+	uint8_t contextId, uint32_t id, uint32_t source, uint32_t target,
 	uint32_t fill, uint8_t flags, const QString &name);
 
 Message
